requirements.txt
google-adk==1.10.0
google-cloud-storage==2.19.0
pdfplumber==0.11.7
jinja2==3.1.6
litellm==1.75.8